RF: #54 split analysis.py into analysis/ module, rename to main.py
