+++ tox.ini
@@ -1,4 +1,4 @@
 [flake8]
-ignore = D100,D101,D102,D103,D104,D105,D200,D201,D204,D205,D400,D401,E266,I100,I201,N802,N803,N806,W503
+ignore = D100,D101,D102,D103,D104,D105,D200,D201,D204,D205,D400,D401,E266,E402,I100,I201,N802,N803,N806,W503
 max-line-length = 120
 exclude = *test*
